requirements.txt
aiohttp==3.6.2
aiounittest==1.4.0
appnope==0.1.0
arrow==0.15.6
astroid==2.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autoflake==1.3.1
autopep8==1.5.3
backcall==0.1.0
bleach==3.1.5
certifi==2020.4.5.1
chardet==3.0.4
colors.py==0.2.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.3.3
entrypoints==0.3
flake8==3.8.3
geonames==0.1.3
googlemaps==4.4.1
howlongtobeatpy==0.1.12
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.32.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numba==0.49.1
numpy==1.18.5
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pyttsx3==2.88
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
timezonefinder==4.4.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.4.2
